wikipedia-scraper/requirements.txt
strands-agents>=1.12.0
boto3>=1.34.0
wikipedia-api>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0

these or higher!